# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "This is a sample API documentation",
        "termsOfService": "http://laravelapi.test/api.terms",
        "contact": {
            "email": "deneme@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravelapi.test/api",
            "description": "Laravel API test Server"
        },
        {
            "url": "http://laravelapi.stage/api",
            "description": "Laravel API Stage Server"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "List all products",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "Limit",
                        "in": "query",
                        "description": "How mant item to return at one time",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paged array of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unathoritized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Create a product",
                "operationId": "store",
                "requestBody": {
                    "description": "store a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Products created response"
                    },
                    "401": {
                        "description": "Unathoritized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/products/{productId}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Info for a psecific product",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the prdoduct to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products detail response"
                    },
                    "401": {
                        "description": "Unathoritized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "Create a product",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the prdoduct to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Products updated response"
                    },
                    "401": {
                        "description": "Unathoritized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Deletes a product",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the prdoduct to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products delete response"
                    },
                    "401": {
                        "description": "Unathoritized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "id column",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductUpdate": {
                "title": "Product Update",
                "description": "Product update model",
                "required": [
                    "name",
                    "slug",
                    "description",
                    "price"
                ],
                "properties": {
                    "price": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "name": "api_token",
                "in": "query"
            },
            "bearer_token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "product",
            "description": "product tag description",
            "externalDocs": {
                "description": "find out more",
                "url": "https://www.tarikunal.com"
            }
        }
    ],
    "externalDocs": {
        "description": "Find out more about Laravel API",
        "url": "https://www.tarikunal.com"
    }
}